event-service/requirements.txt
Flask==2.3.3
pymysql==1.1.0
redis==5.0.1
requests==2.31.0
PyJWT==2.8.0
flask-cors==4.0.1